% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mastodon_accounts.R
\name{get_mastodon_accounts}
\alias{get_mastodon_accounts}
\title{get_mastodon_accounts}
\usage{
get_mastodon_accounts(followers_following)
}
\arguments{
\item{followers_following}{the followers and/or following list of users}
}
\description{
Function to get Mastodon accounts from the "name" field on Twitter.
}
