% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{mastodon_instances}
\alias{mastodon_instances}
\title{utils}
\format{
An object of class \code{character} of length 77.
}
\usage{
mastodon_instances
}
\description{
Utility files
It includes mastodon instancies
}
\keyword{datasets}
