% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migratweet.R
\name{migratweet}
\alias{migratweet}
\title{migratweet}
\usage{
migratweet(user = NULL, followers = TRUE, following = TRUE, academicAPI = NULL)
}
\arguments{
\item{user}{the username on Twitter}

\item{followers}{collect Mastodon accounts from followers. Default to TRUE}

\item{following}{collect Mastodon accounts from following Default to TRUE}

\item{academicAPI}{the Academic Twitter API bearer token. If not specified, retrieved from the from the environmental variable "TWITTER_BEARER" using \link[academictwitteR]{get_bearer}.}
}
\value{
a vector of detected Mastodon accounts and a csv file saved in the working directory that can be imported on Mastodon to add the listed accounts.
}
\description{
Function to collect Mastodon accounts put in the "name" field on Twitter.
}
\details{
Simple function to extract Mastodon accounts from the name field on Twitter using a simple series of regex routines.
}
\examples{
\dontrun{
account_list <- migratweet(user = "NicRighetti")
head(account_list)
}

}
